Starting to refactor email creation
