+++ src/routes/send-invites.js
@@ -1,6 +1,6 @@
 const express = require('express');
-const nodemailer = require('nodemailer');
-const { checkAuthenticated } = require('../auth-helper');
+const sendEmail = require('../utils/send-email');
+const { checkAuthenticated } = require('../utils/auth-helper');
 const { Invite } = require('../models');
 const router = express.Router();
 
@@ -8,14 +8,6 @@ const router = express.Router();
 router.post("/", checkAuthenticated, async (req, res) => {
   try {
     const invites = await Invite.find({});
-
-    const transporter = nodemailer.createTransport({
-      service: 'gmail',
-      auth: {
-            user: process.env.GMAIL_USER,
-            pass: process.env.GMAIL_PASS
-        }
-    });
 
     invites.forEach(function (invite) {
       const url = process.env.HOST + '/invite/' + invite.guid
@@ -27,7 +19,7 @@ router.post("/", checkAuthenticated, async (req, res) => {
           html: invite.names +",<br/><br/>You are invited to our wedding! <a href='" + url + "'>View your invite</a> for the wedding details and use the form on the website to RSVP.<br/><br/> If you have any questions, feel free to reply to this email! <br/><br/>We hope to see you there!<br/>Brittani and Riley"
       };
 
-      transporter.sendMail(message, function(err, info) {});
+      sendEmail(message);
     });
 
     res.json({success : "Success", status : 200});
